# jacd/backend/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "JACD REST API DOC",
    "description": "This document contains the api parameters and responses to be utilised by the frontend."
  },
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "fields": {
      "Email": {
        "type": "string",
        "example": "example@unsw.edu.au",
        "description": "Email address of the user attempting to login"
      },
...
    },
    "parameters": {
      "AuthRegisterBody": {
        "name": "Register info",
        "in": "body",
        "schema": {
          "type": "object",
          "properties": {
            "email": {
              "$ref": "#/components/fields/Email"
            },
            "password": {
              "$ref": "#/components/fields/Password"
            },
            "fname": {
              "$ref": "#/components/fields/FirstName"
            },
            "lname": {
              "$ref": "#/components/fields/LastName"
            }
          }
        }
      },
      "AuthLoginBody": {
        "name": "Login info",
        "in": "body",
        "schema": {
          "type": "object",
          "properties": {
            "email": {
              "$ref": "#/components/fields/Email"
            },
            "password": {
              "$ref": "#/components/fields/Password"
            },
            "rememberMe": {
              "$ref": "#/components/fields/RememberMe"
            }
          }
        }
      },
      "Authorization": {
        "name": "Authorization",
        "in": "header",
        "schema": {
          "type": "string",
          "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImhheWRlbkB1bnN3LmVkdS5hdSIsImlhdCI6MTYwMzk0MzIzMH0.b37PfwlcH_cue6yhgvDt2IiNvhRACf79hTNtacYB94Q"
        },
        "required": true,
        "description": "Bearer [token]"
      },
      "ProjectsCreate": {
        "name": "Create project body",
        "description": "Details required to create new project",
        "in": "body",
        "schema": {
          "type": "object",
          "properties": {
            "u_id": {
              "$ref": "#/components/fields/UserId"
            },
            "p_name": {
              "$ref": "#/components/fields/ProjectName"
            },
            "p_description": {
              "$ref": "#/components/fields/ProjectDescription"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "400": {
        "description": "Bad Input",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "description": "Error message returned from server",
              "example": "Invalid input"
            }
          }
        }
      },
      "403": {
        "description": "Forbidden",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "description": "Error message returned from server",
              "example": "Invalid Token"
            }
          }
        }
      },
      "409": {
        "description": "Conflict",
        "schema": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "description": "Error message returned from server",
              "example": "Project Name Already Exists"
            }
          }
        }
      }
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Given correct user credentials, return an authorised access token to make subsequent user calls",
        "tags": [
          "User Auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AuthLoginBody"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Token"
          },
          "400": {
            "$ref": "#/components/responses/400"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Send registration request for a new user to create an account",
        "tags": [
          "User Auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AuthRegisterBody"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Token"
          },
          "400": {
            "$ref": "#/components/responses/400"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Invalidate a particular authorised token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "User Auth"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/projects": {
      "get": {
        "summary": "Get a list of all projects",
        "tags": [
          "Project Management"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/ProjectsGetAll"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/projects/create": {
      "post": {
        "summary": "Create a new project",
        "tags": [
          "Project Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/ProjectsCreate"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/ProjectsCreate"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        }
      }
    },
    "/projects/{ProjectId}": {
      "delete": {
        "summary": "Delete a particular project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Project Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "get": {
        "summary": "Return all available data of a project",
        "tags": [
          "Project Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/ProjectsGet"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "put": {
        "summary": "Update the details of a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Project Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          },
          {
            "$ref": "#/components/parameters/ProjectsPut"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "409": {
            "$ref": "#/components/responses/409"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get a list of all tasks",
        "tags": [
          "Task Management"
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/TasksGetAll"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/tasks/create/{ProjectId}": {
      "post": {
        "summary": "Create a new task in project",
        "tags": [
          "Task Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          },
          {
            "$ref": "#/components/parameters/TasksCreate"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/TasksCreate"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/tasks/projects/{ProjectId}": {
      "get": {
        "summary": "Get all tasks in project",
        "tags": [
          "Task Management"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/ProjectIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/TasksGetAll"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/tasks/{TaskId}": {
      "delete": {
        "summary": "Delete a particular task",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Task Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/TaskIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "get": {
        "summary": "Return all available data of a task",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Task Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/TasksGet"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "put": {
        "summary": "Update the details of a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Task Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/TaskIdQuery"
          },
          {
            "$ref": "#/components/parameters/TasksPut"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
    "/tasks/user/{UserId}": {
      "get": {
        "summary": "Return all tasks assigned to user id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Task Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/TasksGetAll"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    },
...
    "/users/{UserId}": {
      "get": {
        "summary": "Get details of user",
        "tags": [
          "Profile and User Management"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/UsersGet"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "put": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Profile and User Management"
        ],
        "summary": "Update user's own profile",
        "description": "Updates the user's name, email, password",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdQuery"
          },
          {
            "$ref": "#/components/parameters/UsersPut"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Profile and User Management"
        ],
        "summary": "Delete the user's account",
        "description": "Deletes account and logs out (invalidating token)",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdQuery"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200/Empty"
          },
          "403": {
            "$ref": "#/components/responses/403"
          }
        }
      }
    }
  }
}